adaptada la estructura al nuevo backend
