+++ src/config.js
@@ -1,4 +1,4 @@
 const config = {
-    backendUrl: process.env["BACKEND_URL"] || 'https://peludos-osw-back.herokuapp.com',
+    backendUrl: process.env["BACKEND_URL"] || 'https://peluditos-osw.herokuapp.com',
 };
 export default config;
